+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Changes
 
-## 0.3.4 ->
+## 0.3.4 -> 0.3.5
 
 * Ensure that `Vec3.to_direction()` and `GlobalVector.to_direction()` also work for "Up"
 * Fix `vessel.available_thrust` to calculate the thrust in the current situation (i.e. atmosphere/vacuum)